@@ -31,10 +31,7 @@ from types import MethodDescriptorType, ModuleType
 from IPython.utils.decorators import undoc
 
 
-if sys.version_info < (3, 11):
-    from typing_extensions import Self, LiteralString
-else:
-    from typing import Self, LiteralString
+from typing import Self, LiteralString
 
 if sys.version_info < (3, 12):
     from typing_extensions import TypeAliasType
