@@ -76,14 +76,9 @@ def foo(self, args):
 
 
 def test_magic_arguments():
-    # “optional arguments” was replaced with “options” in argparse help
-    # https://docs.python.org/3/whatsnew/3.10.html#argparse
-    # https://bugs.python.org/issue9694
-    options = "optional arguments" if sys.version_info < (3, 10) else "options"
-
     assert (
         magic_foo1.__doc__
-        == f"::\n\n  %foo1 [-f FOO]\n\n{LEADING_SPACE}A docstring.\n\n{options}:\n  -f FOO, --foo FOO  an argument\n"
+        == f"::\n\n  %foo1 [-f FOO]\n\n{LEADING_SPACE}A docstring.\n\noptions:\n  -f FOO, --foo FOO  an argument\n"
     )
     assert getattr(magic_foo1, "argcmd_name", None) == None
     assert real_name(magic_foo1) == "foo1"
@@ -98,7 +93,7 @@ def test_magic_arguments():
 
     assert (
         magic_foo3.__doc__
-        == f"::\n\n  %foo3 [-f FOO] [-b BAR] [-z BAZ]\n\n{LEADING_SPACE}A docstring.\n\n{options}:\n  -f FOO, --foo FOO  an argument\n\nGroup:\n  -b BAR, --bar BAR  a grouped argument\n\nSecond Group:\n  -z BAZ, --baz BAZ  another grouped argument\n"
+        == f"::\n\n  %foo3 [-f FOO] [-b BAR] [-z BAZ]\n\n{LEADING_SPACE}A docstring.\n\noptions:\n  -f FOO, --foo FOO  an argument\n\nGroup:\n  -b BAR, --bar BAR  a grouped argument\n\nSecond Group:\n  -z BAZ, --baz BAZ  another grouped argument\n"
     )
     assert getattr(magic_foo3, "argcmd_name", None) == None
     assert real_name(magic_foo3) == "foo3"
@@ -107,7 +102,7 @@ def test_magic_arguments():
 
     assert (
         magic_foo4.__doc__
-        == f"::\n\n  %foo4 [-f FOO]\n\n{LEADING_SPACE}A docstring.\n\n{options}:\n  -f FOO, --foo FOO  an argument\n"
+        == f"::\n\n  %foo4 [-f FOO]\n\n{LEADING_SPACE}A docstring.\n\noptions:\n  -f FOO, --foo FOO  an argument\n"
     )
     assert getattr(magic_foo4, "argcmd_name", None) == None
     assert real_name(magic_foo4) == "foo4"
@@ -116,7 +111,7 @@ def test_magic_arguments():
 
     assert (
         magic_foo5.__doc__
-        == f"::\n\n  %frobnicate [-f FOO]\n\n{LEADING_SPACE}A docstring.\n\n{options}:\n  -f FOO, --foo FOO  an argument\n"
+        == f"::\n\n  %frobnicate [-f FOO]\n\n{LEADING_SPACE}A docstring.\n\noptions:\n  -f FOO, --foo FOO  an argument\n"
     )
     assert getattr(magic_foo5, "argcmd_name", None) == "frobnicate"
     assert real_name(magic_foo5) == "frobnicate"
@@ -125,7 +120,7 @@ def test_magic_arguments():
 
     assert (
         magic_magic_foo.__doc__
-        == f"::\n\n  %magic_foo [-f FOO]\n\n{LEADING_SPACE}A docstring.\n\n{options}:\n  -f FOO, --foo FOO  an argument\n"
+        == f"::\n\n  %magic_foo [-f FOO]\n\n{LEADING_SPACE}A docstring.\n\noptions:\n  -f FOO, --foo FOO  an argument\n"
     )
     assert getattr(magic_magic_foo, "argcmd_name", None) == None
     assert real_name(magic_magic_foo) == "magic_foo"
@@ -134,7 +129,7 @@ def test_magic_arguments():
 
     assert (
         foo.__doc__
-        == f"::\n\n  %foo [-f FOO]\n\n{LEADING_SPACE}A docstring.\n\n{options}:\n  -f FOO, --foo FOO  an argument\n"
+        == f"::\n\n  %foo [-f FOO]\n\n{LEADING_SPACE}A docstring.\n\noptions:\n  -f FOO, --foo FOO  an argument\n"
     )
     assert getattr(foo, "argcmd_name", None) == None
     assert real_name(foo) == "foo"
